==> 8. Use Cases/credit-scoring-data-quality/02_Data_profiling.ipynb <==
# 2. Exploratory data analysis - In-depth profiling

The first step in a data preparation pipeline is the exploratory data analysis (EDA). In a nutshell, data exploration and data cleansing are hand-to-hand and both are mutually iterative steps.

*But what does data exploration includes? And how to make a better data exploration giving we are building a credit scorecard model?*


Data exploration includes both univariate and bivariate analysis and ranges from univariate statistics and frequency distributions to correlations, cross-tabulation, and characteristic analysis.
add here detail about pandas-profiling and data exploration in general (re-use the sentence above)

## Read the data & computed metadata

### Import needed packages

import os

from pickle import load
import pandas as pd

from ydata.utils.formats import read_json
from ydata.metadata import Metadata
from ydata.labs.datasources import DataSources

from ydata.profiling import ProfileReport

dataset = DataSources.get(uid='{datasource-id}').read()

try:
    label = os.environ['LABEL_NAME']
except:
    label = 'SeriousDlqin2yrs'

meta = Metadata.load('metadata.pkl')
print(meta)

## Generating the full data profile

try:
    data_path = os.environ['DATASET_PATH']
except:
    data_path = 'train.csv'

try:
    data_split=os.environ['DATA_SPLIT']
except:
    data_split='train'

data = pd.read_csv(data_path)
data = data.drop('Unnamed: 0', axis=1)

from pandas_profiling import ProfileReport

print(f'Profile Name: {data_split}_profile')
profile = ProfileReport(df=data, title='Hotel bookings demand')
profile.config.html.navbar_show = False

profile.to_file(f'{data_split}_profile.html')

## Add here the html artifact to be generated

y = data[label]
ratio_labels = pd.DataFrame(y.value_counts(normalize=True))

print(ratio_labels)

import json

metadata = {
    'outputs' : [
        {
      'type': 'table',
      'storage': 'inline',
      'format': 'csv',
      'header': list(ratio_labels.columns),
      'source': ratio_labels.to_csv(header=False, index=True)
    },
        {
      'type': 'web-app',
      'storage': 'inline',
      'source': profile.to_html(),
    }
    ]
  }

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(metadata, metadata_file)
