==> 8. Use Cases/credit-scoring-data-quality/05_outliers.ipynb <==
# 5. Outliers processing

## Identification & processing

### Import the needed packages

import pickle 

import pandas as pd

from ydata.dataset import Dataset
from ydata.metadata import Metadata

from ydata.quality.outlier import OutlierEngine, OutlierSteps, OutlierPipeline
from ydata.quality.outlier.detection import IsolationForest, StandardDeviation
from ydata.quality.outlier.projection import PCA, ICA
from ydata.quality.outlier.clustering import HDBSCANClustering

prep = pickle.load(open('prep_parameters.pkl', 'rb'))

## Identify the existing outliers

## Read the data from the last pipeline step
train_data = pd.read_csv('prep_traindata.csv', index_col=[0])

train_data = train_data[train_data['age'] != 101]
train_data = train_data[train_data['age'] != 99]
train_data = train_data[train_data['age'] != 0]

## Pipeline outputs

train_data.to_csv('prep_traindata.csv')

pickle.dump(prep, open('prep_parameters.pkl', 'wb'))
